feat: Introduce anonymous, opt-in telemetry via Sentry
This commit introduces anonymous, opt-in telemetry which
collects crush dumps and sends it to Sentry.  The default
value of the new configuration option is false/"off":

  - env var: `KRAFTKIT_COLLECT_ANONYMOUS_TELEMETRY`
  - config file: `collect_anonymous_telemetry`
  - CLI flag: `--collect-anonymous-telemetry`

We're introducing the option in order to improve the
CLI experience for endusers and tackle panics and handle
bug reports in a more systematic and prompt manner.

The local build experience totally disables telemetry
and it's only included as part of releases.  This is
done because 1). contributors are making iterative updates
may encounter higher error-rates during developmenet and
2). local-first development may include changes which
are not "official" and therefore not debuggable from
a telemetry point-of-view.

Signed-off-by: Alexander Jung <alex@unikraft.io>