feat(pkg): Support packaging without kernel
This commit introduces the ability to package a runtime Kraftfile
without an actual kernel.  This is a nuanced and undefined
behaviour but can be useful in some situations (like noops).  It
also lays down the ground in the future to be able to handle
packaging with other contents where a kernel is implicit to
a remote system.

Signed-off-by: Alexander Jung <alex@unikraft.io>