feat: Add support for auxiliary read-only memory blob files
This commit introduces initial support for packaging auxiliary
read-only memory blob files (auxiliary ROMs).  These are totally
arbitrary files which can be used by an external provider and
mounted to the kernel.  It's up to the application to decide
what to do with this and provides flexibility in terms of being
able to support additional types of usecases beyond a single
initramfs.

Right now this can be used in two different ways:

1. Via a `Kraftfile`:
   ```yaml
   spec: v0.6

   runtime: base:latest

   rootfs: ./rootfs

   roms:
   - path/to/blob.bin
   ```

2. Via CLI:
   ```
   kraft pkg \
     --name test:latest \
     --rom path/to/blob.bin \
     [...]
   ```

Signed-off-by: Alexander Jung <alex@unikraft.io>