feat: Add `Export` method to package interface
This commit introduces an `Export` to the package interface.  As
a result, `Export` is added to all implementing structs.

The main implementing package (manager) is OCI.  This will export
the package into an OCI archive format[0].  This can be used
to transport the package as a tarball, and import it into
OCI-compatible registries or tools, e.g. `docker import`.

[0]: https://buildpacks.io/docs/for-app-developers/how-to/special-cases/export-to-oci-layout/

Signed-off-by: Alexander Jung <alex@unikraft.io>