refactor(oci)!: Use the manifest digest instead of index digest as ID
Since within KraftKit and the OCI package manager, a package
refers to an index and a single manifest, and never a group
(even though we can package multiple packages into an index),
use the manifest digest as the ID and not the index.  This makes
the OCI package ID properly unique.

Signed-off-by: Alexander Jung <alex@unikraft.io>