fix(oci): Indent index manifest JSON objects with newline
This is a bit of a counter-intuitive change.  The reason for this is
that the remote OCI registry returns index manifests in this format.
Other digest-ables, like image config, and image manifest, are not
indented; so only the index is indented.  This is important change
bceause it affects the digest of the index, which is important to
try and maintain (when no changes really occur) so that comparisons
and lookups can be made effectively.

Signed-off-by: Alexander Jung <alex@unikraft.com>