fix(fs): Add fail-early check for conversion
We do not support cpio->erofs and erofs->cpio builds
so we need to fail early on this. Previously, this
was not caught as erofs has a valid tar start and
it would be caught by that.

We also can't use the full check in the respective
packages as it would cause a circular dependency.

This simple check only looks for the magic
number and does not do any other advanced parsing

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>