feat: use docker builder if available
Docker includes a buildkit instance that we can use. We do need the
containerd-storage enabled version (sad), but this is now the default,
see https://docs.docker.com/engine/storage/containerd/.

This is *so* much more performant, is always available if docker is
running, and never has weird conflicts where it can't start.

Signed-off-by: Justin Chadwell <justin@unikraft.com>